move CompareCommitStateDB to test file
